projects/9/django_template/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-grappelli==3.0.8
djangorestframework==3.15.1
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.7
openpyxl==3.1.2
packaging==24.0
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
